FixSymmetry checks if clause is in set
The regression of a few tests between revisions
r2432 and r2433 are due to the fact that the
fixSymmetry method is senstitive to the order in
which the clause set is presented. This commit
introduces a check: before trying to find
symmetric clauses, check if the clause itself is
not part of the clause set.

Also organizing the code Robinson2LK and removing
an unused debug function.

git-svn-id: http://gapt.googlecode.com/svn/trunk@2445 85d21ef8-add0-11de-ae77-5bfb9c5a3f34