Fixed a bug in the parsing of prover9 formulas.
The logical constants where parsed unsoundly: the top constant was
parsed as an unsatisfiable formula, and the bottom constant as a valid
formula, instead of vice versa.

This bug is a reminder that we should at some point discuss how we
should handle the logical constants top and bottom in GAPT - I think
some algorithms do not handle them correctly at the moment (this is,
apparantly, the reason that they are not parsed "directly").