Started work: CutIntro proof-building for >1 cuts
Following the theoretical description in "Algorithmic Introduction of
Quantified Cuts (Hetzl et al 2013)", we generalize the current code
(which builds a proof with one cut according to the paper) to building a
proof with n cuts (as determined by the extended Herbrand sequent
given). We try to keep the algorithm close to the theoretical
description; the main deviation is that, instead of building a proof
with implication-left and and-left rules which are converted to cut and
forall-left rules in a second step, we directly build the proof with
cuts and forall-left rules.

For sake of simplicity, we leave out the use of interpolants in the
construction (i.e. we use A_i instead of A_i'). This missing feature
should be implemented in the future.