Simplifying computation of canonical solution
For aesthetic purposes, the alpha variables from
the grammar were replaced by x variables before
the terms were used in the cut formula. This is
only because we are used to seeing quantifications
over x and not alpha. But the use of more complex
grammars makes this replacement more complicated.
This commit removes this complexity from the
computation of canonical solutions by using the
terms 'as is' and quantifying over alpha
variables.