changed expansion tree toSequent to toFSequent
adjusted the interface of cut-introduction to use an FSequent (a consequence of changing the expansion sequent toFSequent method). can you check if this change is ok for you? the cli interface still has both Sequent and FSequent arguments for compatibility with scripts.

fixed a recently introduced bug in thf export