Changed symbol renaming to a member method
I added methods "renameSymbols" to {HOL,FOL}{Expression,Formula}
that take a SymbolMap as an argument. These methods are intended
as replacements for the various apply methods of the NameReplacement
class in algorithms/rewriting, with the goal of making this class
obsolete. Note that I haven't dealt with resolution proofs yet.