Log the construction time of the proof chosen
When many grammars are found in the
cut-introduction algorithm (delta-table method),
the proofs for all of them are constructed and the
smallest one is chosen. The time logged for this
operation was the accumulated time for
constructing all of the proofs. For the sake of
comparisson, this commit changes the behaviour so
that only the time taken to construct the smallest
proof is logged.