Generates symmetry instances for the input formula
Issue #300 was due to the fact that no symmetry
instances were generated for the formula which was
being proven, but only for the equality axioms
used. This commit fixes this issue by collecting
the equalities occurring in the input formulas and
generating symmetry instances for them.

As a result, one test in MistTest fails with stack
overflow (since the number of formulas increased).