started an llk output for ral proofs.
fixing ntape example scripts

introduced a file which collects different formatters (at the moment toPrettyString, the LLK formula renderer and theLLK Latex renderer). Also cleaned up the implementation of toPrettyString, even though after the refactoring, it is only used once in prooftool's draw sequent. It might still be useful in the future - perhaps we can define an object Formatter with methods for each base element (expr, type, perhaps fsequent) of a language which can be changed via a command-line parameter.