leanCoP import: magical dnf transformation
Different formula trees will not match, even if
they are indeed the same formula (e.g. A v (B v C)
does not match (A v B) v C). This is a problem
when trying to find the substitutions for leanCoP
formulas. The implemented magical DNF tries to
match the structure parsed from the files.