Fix race condition in withTimeout.
Sometimes in the regression tests, a ThreadDeath exception was caught by
specs2.  This happened because upon Thread.stop(), we 1) caught the
ThreadDeath exception in the worker thread, 2) then recorded that
exception in the result variable, and 2+-ε) read the result variable in
the main thread at some undefined point before or after 2).

This change waits for the worker thread to finish before reading result.

In addition, withTimeout now takes a Duration parameter (instead of
long), making it clear what unit the timeout is in; the ThreadDeath
exception is recorded in the TimeOutException as well (this shows where
the timeout actually occurred).