veriT interface: constants vs variables
Fixed issue #304.
Prints variables with a '$' sign at the end for
veriT and parser those symbols as variables
afterwards. Note that in order to get the
substitutions we need to change x and y to :x and
:y, since FOLMatchingAlgorithm does not return a
substitution from a variable to itself (and it
should not).

This solution is less elegant than a table, but it
works also if the export/parsing is done outside
veriTProver.