bugfix in ExpansionProofToLKProof
This fixes a bug which occured in the presence of
a quantified formula with several instances of
which some are used for proving the goal sequent
and some are not. The contraction is now only
carried out if the main formula exists both in the
previous context and is created by the current
inference.