VeriT symmetry instances: addSymmetry algorithm
Implements an addSymmetry algorithm which should
be used as a post-processing step after parsing a
veriT proof. It is run automatically after using
VeriTProver to prove a sequent, but not after
parsing. Therefore, the result of an arbitrary
parsed veriT proof is a quasi-tatutological
expansion sequent, while the result of using
VeriTProver is a tautological expansion sequent.