cleaning up utility methods
This commit pulls utility methods which work on
terms and formulas from various places in the
project into the expr-package (into utils.scala,
hol/utils.scala and fol/utils.scala). Furthermore
it uniformizes them, removes duplicates, adds
scaladoc and some unit tests.

As part of this I have changed the freeVariables-
object to return a set of variables and not a list
(the previously implemented algorithm based on
lists was quadratic, this one is n log n).