add weakening nodes to expansion trees
This provides a clean way of specifying that a
certain part of an expansion tree is not needed
for obtaining a tautology (or, respectively, a
quasi-tautology, etc.). A weakening node n carries
a formula f. The shallow formula of n is f, the
deep formula of n is top if n has negative
polarity and bottom if n has positive polarity.

Fixes #328 and #340.