Added more control over how tactics are applied
I added a new type TacticApplyMode that can be one of three things:

*) An "OnLabel(label: String)",
*) a  "UniqueFormula", or
*) an "AnyFormula".

This type replaces Option[String] in tactics constructors and determines how the tactic behaves. OnLabel(l) means that the tactic will be applied to the formula with label l and fail if that's not possible (aka the old Some(l)). UniqueFormula means that the tactic will succeed iff there is exactly one formula it can be applied to. AnyFormula is the old "None" behavior, i.e. the tactic succeeds iff any formula can be found that it can be applied to.

The good part is that AnyFormula is not user-facing: it is not possible to use this mode via the functions defined in gaptic/package.scala. The reason it exists is that it's convenient for defining bigger tactics like decompose.