Babel parser uses signatures
How it works: If you do nothing, the Babel parser will handle constants and variables as before, i.e. identifiers beginning with (upper- or lowercase) u,v,w,x,y,z are considered variables, the rest are constants. But if the parser finds an implicit value of type Signature in the local scope, then that will be used instead. Using the signature class, you can control exactly which identifiers should denote variables and which should be constants, and also what their types should be. The most common use case is captured by the MapSignature class: You supply a map of type String => Type that contains exactly the names of constants. This means that a name stands for a constant if the map is defined on it and for a variable otherwise.