Changed behavior of UnfoldTactic
The tactic still accepts a list of definitions, but now it will only eliminate the first one in the list that actually occurs in the formula and leave the rest. If none of the supplied definitions occur, it will fail. This means that you can still access the old behavior by using repeat.