Unified equality tactics
I saw no reason why there have to be 2 equality tactics, so now there's only one. Moreover, if you do not specify a direction or target formula, then the direction has to be unambiguous. This means that if you try to apply the equation s=t to a formula containing both s and t, the tactic will fail.