Hygienic term replacements.
Modern first-order provers such as Vampire introduce new predicate
symbols in a refutation--even if they start from a CNF.

For example, Vampire likes to introduce the nullary predicate symbol
$spl0 during splitting.  What happens when the end-sequent in GAPT
already contains the name "$spl0"?  During export, we rename it to e.g.
"f37", and after proof import we rename it back to "$spl0"--and now we
have a clash: "$spl0" is both the defined predicate for splitting and a
predicate that occurs in the end-sequent.

The solution is obviously to replace "$spl0" by "$spl0_0" when renaming
back the proof, which is precisely what we do here.