[WIP] Simplifying prooftool structure
There is now an abstract subclass of ProofToolViewer for scrollable
prooftool windows. The removal of the scroll pane from ProofToolViewer,
which ExpansionSequentViewer extends, means that the preferred size of
the DrawSequentProof object can't be computed in the same way as before.
Therefore, the ExpansionSequentViewer is currently somewhat broken. I
haven't figured out yet how to properly compute this size.