LKToExpansionProof: handle EqualityLeftRule correctly
When extracting an expansion proof, we need to modify the trees for the
equations, this is necessary if they are introduced by weakenings.  We
first removed the equation, and then modified the equation at the
original(!) index.  This failed if the equation was on the right side of
the auxiliary formula.

I honestly have no idea why we only hit this bug now.