Use logging instead of verbose flags.
There are a few places where we want to output progress and/or more
detailed information.  For example in cut-introduction, we passed around
a verbose flag.  Since we already use logging, I changed all of the
progress/detail outputs to use logging instead, and added a small method
to enable logging at runtime for a specific class:

  CutIntroduction.makeVerbose()