cut-intro: use many-sorted instance terms
Cut-introduction is still first-order.  This is only a small technical
change: instead of using first-order terms to encode the instances of
the Herbrand sequents, we now use a differently-sorted root symbol.

As a result, *every* VTRATG with first-order non-terminals (apart from
the start symbol) that covers the language corresponds to a proof.  The
following restriction is now unnecessary:

  productions of the form "τ → α" are disallowed