Prooftool: Lines now align correctly
Alright, this is a big one. The math for figuring out the correct
alignments is actually a bit involved, but I seem to finally have gotten
it right. There's also much less manual calculation of widths, which
makes the whole thing more robust (turns out that unlike FlowPanels,
BoxPanels are very well-behaved when it comes to their dimensions).

I also changed a few other things along the way:
*) The popup menu now opens at the correct position.
*) The popup menu now contains a "display subproof" option that opens a
subproof in a new window.
*) If a subproof is collapsed, you can click on the "(...)" label to
expand it.
*) Scrolling now works correctly.
*) Removed the blue highlighting of sequents because it interfered with
the highlighting of individual formulas.