Prooftool: Added marking functionality
You can now right-click on any formula in an LK proof and choose
to mark its ancestors, descendants, or both. All of these options
include the formula itself. The marked formulas will be displayed
with a green background. Currently, the only way to get rid of
the markings again is via the "Remove all markings" button in the
view menu.