Added sequents for NJ calculus
NJSequent is a separate type from Sequent, but there is an
implicit conversion. This means that even though NJSequent
has no methods of its own, you can call any method that
exists on Sequent on an NJSequent and it will work by
converting to a Sequent first.

This is obviously not what we want in every situation: suppose you
have an NJSequent s. It is reasonable that s :+ A should yield
a Sequent, since the result is not a valid NJSequent, but there
is no good reason that A +: s shouldn't be an NJSequent. This
means that some methods will need to be reimplemented on NJSequent.