Refactor Context and expression-level reduction.
See the scaladoc for Context for details about the implementation.  This
commit changes contexts in two significant aspects:

1) There is now a clear and transparent distinction between the
elements/modifications to a Context (e.g. adding a defined constant),
and the effect they produce (adding the constant, and adding the
definition).  The modifications are stored as Elements as before, the
state is now stored in an extensible record comprised of facets.

2) The Context is also much more extensible: you can add new kinds of
information without changing the main class.

While I was at it, I also changed the expression-level normalization
code (previously BetaReduction):
3) Normalization is now parameterized by (potentially more than one)
reduction rule.  Definitional reduction is now handled at the same as
beta-reduction.

=> It is now clear how to extend Context to store LK proofs.  (Just add
a facet, and an element to add them).

=> Adding support for projections and the associated reduction rules is
straightforward (one approach is to extend the
StructurallyInductiveTypes facet to include information about the
projections, and make it implement ReductionRule so that projections get
reduced).

The main question here is how to integrate this in LK, one possibility
is to change the definition rules to general equivalence rules (i.e.
equal after reduction).

=> Recursors are straightforward to implement as well (see projections).

=> We can add primitive recursive functions that definitially reduce as
well.