Working on Proof schema instantiation.
Wrote a function which commputes the depth of a lambda term. The point of this
function is to test when a set of arguments to a given proof might cause an
infinite tree. The basic function for testing this case multiplies the
term depths of all input args. Also, I will make a function for instantiating
proof with numeric parameters and one for instantiating proofs with any
inductive data type.