SkolemFunctions: do not verify acyclicity
This fixes a performance problem when clausifying formulas with many
quantifiers such as `?x1 ... ?x100 p(x1, ..., x100)`.  Previously, we
would compute the dependency order with every additional Skolem
function, resulting in cubic runtime complexity.

We still check acyclicity via the Context: whenever a new Skolem
function is added, we verify that its definition only depends on
previously declared constants.

In fact the check in the SkolemFunctions was not even sufficient in the
case of definitions.  With definitions, you could create cyclic Skolem
definitions that would not have been detected by the old check:
  s_1 -> ∃x P(x)
  s_2 -> ∃x Q(x, s_1)
  P(x) := Q(x, s_2)