LKToExpansionProof: add heuristic for intuitistic proofs
Essentially the only change is that negation-left and implication-left
rules are tried *last*.  LKToExpansionProof now constructs proofs in LJ
for a large and important class of sequents, having the following form:

  ∀C_1, ..., ∀C_n ⊢ A

Where C_i are clauses (disjunctions of literals), and A is a cube
(conjunction of literals).  It might also work if the C_i are
implications, but I haven't thought about it carefully.