update sbt-scoverage plugin / remove scalariform
the previous sbt-scoverage plugin version seems to be incompatible with newest sbt and scala version.
furthermore, the latest version of sbt-scoverage has a dependency conflict with the latest scalariform version (which is already over 4 years old).
therefore we remove scalariform for now and add an up-to-date formatter at a later point