make gaptic macro compatible with scala 3
we now require that implementations of LemmaHelper implement handleTacticBlock without implicit arguments.
instead, implicit arguments should now be passed via a custom inline apply method on the implementing class' companion object which in turn passes them to the constructor and calls the macro with the appropriate LemmaHelper.

the reasons for this change are:
- it more clearly communicates how and where implicit arguments are passed since now the inline apply method exactly specifies the required implicit arguments
- it improves type safety: by requiring a handleTacticBlock implementation the compiler can guarantee that the implementations have this exact method
- it keeps the macro implementation simple as in the scala 3 macro system there is no straightforward way to call a method on a trait that might not have this required method implemented without resorting to low-level AST operations and somewhat reimplementing the expected scala semantics
- it makes it more transparent which apply method is called when using an implementation of LemmaHelper
- it allows to create functional implementations of LemmaHelper

this requires now that implementations of LemmaHelper get the implicit arguments not from the handleTacticBlock method but the surrounding context, for example from constructor parameters.

the apply method for LemmaHelper is renamed to applyTactics and now serves as a building block for the custom apply methods of implementations.

for now we also separate the old scala 2 implementations of LemmaHelper from the scala 3 implementations as with the new design we need to make the apply methods inline (which is not available in scala 2) so that the macro code can do its job. this allows us to keep checking the old scala 2 implementation if there are bugs in the implementation of the scala 3 macro